requirements.txt
torch~=2.0.0
torchvision
diffusers~=0.16.1
transformers
urllib3>=1.26.15
dingtalk-stream>=0.1.0
Pillow~=9.5.0